Add blip-diff for basic patch generation.
Because it doesn't matter exactly how blip-diff works as long as it
generates a proper patch, there's not much of a unit-test suite for it.
However, there is a blip-diff-test tool, which takes two filenames,
generates a diff, then applies that diff in-memory and checks after each
operation than the original target and the reconstructed target are
identical. That should help catch a few bugs.