Add classes representing blip patch operations.
The existing python-blip codebase is littered with things that dispatch
on the type of a blip operation and do various calculations as a result;
it makes a lot more sense to have a central class representing such
concepts, with the various calculations implemented as methods.

These operation classes aren't hooked into the rest of the codebase yet,
but they're implemented and their tests pass.