Batch up consecutive TargetRead operations.
Rather than issue a bunch of one-byte TargetRead operations, when we
can't encode a particular target block we just increment a counter. Now
iter_candidate_ops() will include the TargetRead with each candidate
operation, so that op_efficiency() can assess the actual operations that
will be encoded.

Since 'merging consecutive TargetRead operations' was the only thing
that the optimize pass was doing, I've removed it from the processing
chain in bin/bps-diff. It's still a standalone tool in bin/bps-optimize
though.