Implement backwards block extension.
Because the blocks we use for chunking the source and target files do
not overlap, it's possible that a change from the source to the target
file will not be block-aligned, and we'll produce up to a block's worth
of TargetRead operations for a single byte change.

Now that we don't produce TargetRead operations immediately, when we do
find a match we can extend it backwards - in the best case, the new
match will extend backwards all the way to the beginning of the
potential TargetRead, and we won't have to produce a TargetRead at all.

This costs CPU time as we measure matches backwards as well as forwards,
but we also now generate three Python objects per match (a TargetRead
op, the Copy operation we're interested in, and a list to group them
together), and I suspect this is what makes this code twice as slow as
before.