Search harder for block encoding candidates.
When we slice the source and target files up into blocks, each block
begins at an integer multiple of the block size, but there's no
guarantee that the changes between the source and target files will be
aligned to block boundaries. If we're trying to encode a block beginning
at a particular offset in the target file and there are no matching
blocks anywhere, that's no problem - we can just step forward with
TargetRead operations until we're realigned, and continue happily.
However, if we're trying to encode a block at a particular offset and we
do find a match, previously we would just take it even if skipping
forward a little would have allowed us to find a much better match.

Now, when we're encoding a block at a particular offset, we also try
adding up to another blocksize bytes of TargetRead in case we find
better candidates. Since we now support extending Copy instructions
backwards, these extra potential TargetRead instructions can often be
entirely removed.

The upside of this change is that the resulting patches can be as little
as two thirds the size of patches from previous versions. The downside
is that the current implementation is something like ten times as slow
as previous versions.