Don't call len() in an inner loop.
Profiling revealed that a decent chunk of bps-diff's runtime was spent
calling len(), and the profile was more than twice as fast with calls to
len() moved outside the loop. It doesn't provide nearly as dramatic
a speedup outside the profiler (the bsnes test-case takes maybe 5/6ths
as long as it did before) but even that is more than welcome.