Optimise the measure_op() function.
Profiling tells me that a large chunk of time is spent in the
measure_op() function, so I've shuffled things around so that as few
Python operations as possible are done in the "measure backward" and
"measure forward" loops.

Also, I changed the relationships of various offset variables in the
main loop. Now measure_op()'s "targetoffset" parameter is always the
offset into the target where the block in question was found, and
'pendingTargetReadBytes' extends backwards from that point.