Turn BlockMap into a custom hashmap.
The standard Python dict class keeps a copy of all the keys inserted
into it, presumably so it can resolve hash collisions. Unfortunately,
that means we wind up storing two entire copies of the source and target
files in memory - a contiguous copy we're processing, and the same data
split into fragments and stored in a random order in the dictionary.

For our purposes, we don't care about hash collisions... or rather, now
that measure_op() has learned how to measure an op for which the copy
instruction doesn't match at all, and now that op_efficiency() has
learned to calculate the efficiency of an empty operation, we don't care
about hash collisions. We really want to be able to record the offsets
of as many blocks as possible, so hacking up our own hashmap
implementation makes sense.