Add a new OpBuffer container.
This class is basically a list of operations, except that when adding
a new operation you can include a 'rollback' value, which causes the new
operation to replace or trim the trailing operations in the buffer. So,
if you find a few half-decent matches, then later find a much better
match, you can throw away the mediocre ones and just use the best one.

This means we can make much more efficient patches using matches we
stumble across to greatest advantage, instead of being stuck with poor
matches we've already issued.