Sample every byte of the source and target.
This gives us a better chance at finding The Best Possible Encoding for
a particular offset, at the expense of having to trawl through a lot
more sub-optimal encodings. Now that the BlockMap doesn't store a copy
of every block it maps, we can do this without using a ridiculous amount
of memory.