Add OpBuffer.copy_offsets()
Because of the way the bps file-format works, the byte pattern required
to encode a particular SourceCopy or TargetCopy operation depends on the
offset used by the previous SourceCopy or TargetCopy operation, which
means that if you want to choose the patch operation with the most
efficient encoding, you need to keep track of old copy offsets.

Since we now have an OpBuffer of patch operations, it seems best to let
it keep track of such things.