Ignore inefficient candidate copy operations.
A copy operation (SourceCopy, TargetCopy or SourceRead) is good because
it can represent a large number of bytes in the target file, using only
a small number of bytes in the patch file - as the number of target
bytes increases, the efficiency approaches infinity. On the other hand,
a TargetRead instruction stores the bytes it represents plus overhead,
so as the number of target bytes increases, the efficiency approaches
1 (from below).

While most copy operations have very high efficiency, some do not.
Previously. we always used a copy operation whenever we could - but it
turns out that sometimes, the only available copy operation has
terrible efficiency - such as using 4 patch bytes to encode three target
bytes, for an efficiency of 0.75. Assuming there's already at least
2 pending TargetRead bytes, adding another byte to the TargetRead is
just as efficient as the copy - and if there's more than 2 pending
bytes, adding another is *more* efficient than the copy.

There's probably some clever calculation I could make that would exactly
quantify whether the copy or the TargetRead was more efficient in any
particular instance; for the moment, just plain ignoring any Copy
operation with an efficiency less than 1.0 gives pretty great results.