Make OpBuffer cache copy offsets.
Previously we spent a lot of time scanning through the opbuffer trying
to find copy offsets at some particular rollback distance. Now we cache
the copy offsets after each operation with that operation, so we never
have to go searching for them.

This makes the bsnes test take ~354 seconds in the profiler instead of
~672. A palpable win!