A way to calculate aggregate operation efficiency.
Focussing on the efficiency of individual patch operations can sometimes
lead to "penny wise, pound foolish" situations where a slight efficiency
increase on one operation prevents a major efficiency increase on
another.

Now that we have a way to calculate the efficiency of a sequence of
patch operations, we can consider the bigger picture and make better
encoding decisions.