Handle partial rollbacks more cleverly.
Previously, if an operation was added to the OpBuffer with a rollback
value that reached part-way through a previous operation, we truncated
the previous operation if it was a TargetRead (since that would always
be an efficiency win) and chopped the front off the new operation
otherwise. This was simple, but not always efficient.

Now OpBuffer tests out a number of different approaches, and chooses
whichever one gives the most efficient result.